Añadido live_render en lugar de live_component
