+++ lib/siwapp/invoices.ex
@@ -22,8 +22,11 @@ defmodule Siwapp.Invoices do
   end
 
   def scroll_listing(page, per_page \\ 20) do
-    Query.scroll_list_query(page, per_page) |> Repo.all()
+    Invoice
+    |> Query.scroll_list_query(page, per_page)
+    |> Repo.all()
   end
+
 
   @doc """
   Gets a list on the invoices that match with the params
@@ -112,13 +115,16 @@ defmodule Siwapp.Invoices do
     Invoice.changeset(invoice, attrs)
   end
 
-  def list_past_due_or_pending do
-    Repo.all(Query.list_past_due_or_pending())
+  def list_past_due_or_pending(page, per_page \\ 20) do
+    Invoice
+    |> Query.list_past_due_or_pending()
+    |> Query.scroll_list_query(page, per_page)
+    |> Repo.all()
   end
 
-  def list_past_due do
-    list = list_past_due_or_pending()
-    Enum.filter(list, fn x -> status(x) == :past_due end)
+  def list_past_due(page, per_page \\ 20) do
+    list_past_due_or_pending(page, per_page)
+    |> Enum.filter(fn x -> status(x) == :past_due end)
   end
 
   @spec status(Invoice.t()) :: :draft | :failed | :paid | :pending | :past_due
