+++ lib/siwapp_web/live/invoices_live/invoices_list.html.heex
@@ -1,3 +1,19 @@
+<nav class={"navbar is-fixed-bottom
+             #{if Enum.empty?(@checked), do: "is-invisible"}"}
+            role="navigation" aria-label="main navigation">
+  <div class="navbar-menu">
+    <div class="navbar-end">
+      <div class="navbar-item buttons">
+        <button class="button is-danger">Delete</button>
+        <button class="button is-info">Send Email</button>
+        <button class="button is-info">Set Paid</button>
+        <button class="button is-info">Download PDF</button>
+        <button class="button is-info">Duplicate</button>
+      </div>
+    </div>
+  </div>
+</nav>
+
 <table class="table is-fullwidth is-striped is-bordered is-hoverable">
   <thead>
     <tr>
@@ -54,4 +70,3 @@
   </tbody>
 </table>
 <div id="infinite-scroll" phx-hook="InfiniteScroll" data-page={@page}></div>
-
