--- lib/siwapp_web/live/invoices_live/table_component.ex
@@ -1,5 +1,0 @@
-defmodule SiwappWeb.InvoicesLive.TableComponent do
-  @moduledoc false
-  use SiwappWeb, :live_component
-  alias Siwapp.Invoices
-end
