+++ lib/siwapp_web/live/page_live/index.ex
@@ -6,23 +6,6 @@ defmodule SiwappWeb.PageLive.Index do
   alias Siwapp.Invoices
   alias SiwappWeb.GraphicHelpers
 
-  def mount(_params, _session, socket) do
-    {:ok,
-     socket
-     |> assign(:invoices, Invoices.list_past_due())
-     |> assign(:checked, MapSet.new())}
-  end
-
-  def handle_event("click_checkbox", params, socket) do
-    checked = update_checked(params, socket)
-
-    {:noreply, assign(socket, checked: checked)}
-  end
-
-  def handle_event("edit", %{"id" => id}, socket) do
-    {:noreply, push_redirect(socket, to: Routes.invoices_edit_path(socket, :edit, id))}
-  end
-
   @doc """
   Creates the dashboard chart with information about invoices (amounts of money per day).
   """
@@ -37,25 +20,5 @@ defmodule SiwappWeb.PageLive.Index do
   @spec date_to_naive_type({Date.t(), integer()}) :: {NaiveDateTime.t(), integer()}
   defp date_to_naive_type({date, amount}) do
     {NaiveDateTime.new!(date, ~T[00:00:00]), amount}
-  end
-
-  defp update_checked(%{"id" => "0", "value" => "on"}, socket) do
-    socket.assigns.invoices
-    |> MapSet.new(& &1.id)
-    |> MapSet.put(0)
-  end
-
-  defp update_checked(%{"id" => "0"}, _) do
-    MapSet.new()
-  end
-
-  defp update_checked(%{"id" => id, "value" => "on"}, socket) do
-    MapSet.put(socket.assigns.checked, String.to_integer(id))
-  end
-
-  defp update_checked(%{"id" => id}, socket) do
-    socket.assigns.checked
-    |> MapSet.delete(String.to_integer(id))
-    |> MapSet.delete(0)
   end
 end
